requirements.txt
Cerberus>=1.3,<1.4
devdeck-core>=1.0,<1.1
obs-websocket-py>=0.5,<0.6
